changed rating parameter to float

+++ TravelApp/app/src/main/java/jwest/android_class/travel_app/Landmark.kt
@@ -1,6 +1,6 @@
 package jwest.android_class.travel_app
 
-class Landmark (id : String, authorId : String?, coordinates: Coordinates, title : String, description: String, rating: Int) {
+class Landmark (id : String, authorId : String?, coordinates: Coordinates, title : String, description: String, rating: Float) {
     val id = id
     val authorId = authorId
     var coordinates = coordinates
@@ -8,7 +8,7 @@ class Landmark (id : String, authorId : String?, coordinates: Coordinates, title
     val description = description
     val rating = rating
 
-    constructor() : this("", "", Coordinates(100.0, 100.0), "", "", 0){
+    constructor() : this("", "", Coordinates(100.0, 100.0), "", "", 0f){
 
     }
 }
